Allow passing in nixpkgs and macos sdk
This allows people to more easily use the upstream nixcrpkgs if they
wish by fetching the macos sdk separately (such as from a private s3
bucket), and then passing it as an argument to nixcrpkgs.

For good measure, this also makes it possible to pass in 'nixpkgs' so
that people who are using this from nix can avoid an extra nixpkgs
evaluation.