macos/sdk_builder: use plist, not json file
Not all macos SDKs seem to bundle SDKSettings.json. Some only seem to
have SDKSettings.plist.

Parsing the plist file without ruby is pretty easy, so it seems
straightforward to switch over to that.

In addition, I removed the `ls` check at the end because it didn't seem
to be correct for the SDK I'm using.

Specifically, with my 10.10 sdk, the path seemed to be
'/usr/include/c++/v1/iterator' (note the v1).

However, since everything still worked without the check as best I can
tell, it seems reasonable to just drop it entirely.